requirements.txt
mysql-connector-python==8.0.25
optparse-pretty==0.1.1
protobuf==3.16.0
six==1.16.0
